pincode typecast to integer

+++ src/main/java/com/bharatpe/lending/service/LendingUploadService.java
@@ -82,7 +82,7 @@ public class LendingUploadService {
 		lendingApplication.setStreetAddress(String.valueOf(shopDetails.get("street_address")));
 		lendingApplication.setArea(String.valueOf(shopDetails.get("area")));
 		lendingApplication.setLandmark(String.valueOf(shopDetails.get("landmark")));
-		lendingApplication.setPincode((Long) shopDetails.get("pincode"));
+		lendingApplication.setPincode(((Integer) shopDetails.get("pincode")).longValue());
 		lendingApplication.setCity(String.valueOf(shopDetails.get("city")));
 		lendingApplication.setState(String.valueOf(shopDetails.get("state")));
 		
@@ -147,7 +147,7 @@ public class LendingUploadService {
 		data.setStreetAddress(String.valueOf(shopDetails.get("street_address")));
 		data.setArea(String.valueOf(shopDetails.get("area")));
 		data.setLandmark(String.valueOf(shopDetails.get("landmark")));
-		data.setPincode((Long)(shopDetails.get("pincode")));
+		data.setPincode(((Integer)(shopDetails.get("pincode"))).longValue());
 		data.setCity(String.valueOf(shopDetails.get("city")));
 		data.setState(String.valueOf(shopDetails.get("state")));
